Tidy up out of date documentation
Almost all of this has been superceded by official design docs which,
as part of pulumi/home#29 will be moving to pulumi/design-notes.

Those that aren't will be done as new design docs in the future, since
most of the remainder are vastly out of date and/or aspirational at best.

Resolves pulumi/pulumi#408.